Task: move index to route folder
